iio: mpu: Add Android sensors API 1.3 support.
Bug 1581694
Bug 200060572

Reviewed-on: http://git-master/r/663646
(cherry picked from commit 5f5dfe5f020e2eb8773e012d57e184fadaeb8b06)
Reviewed-on: http://git-master/r/662799

+++ drivers/iio/common/nvs/nvs_iio.c
@@ -938,7 +938,7 @@ static const struct iio_chan_spec nvs_ch_ts[] = {
 static int nvs_chan(struct iio_dev *indio_dev)
 {
 	struct nvs_state *st = iio_priv(indio_dev);
-	size_t buf_sz = 0;
+	unsigned int buf_sz = 0;
 	unsigned int n;
 	unsigned int i;
 	int scan_index;
@@ -1016,7 +1016,7 @@ static int nvs_chan(struct iio_dev *indio_dev)
 	if (scan_index <= 1)
 		return -ENODEV;
 
-	st->buf = devm_kzalloc(st->dev, buf_sz, GFP_KERNEL);
+	st->buf = devm_kzalloc(st->dev, (size_t)buf_sz, GFP_KERNEL);
 	if (st->buf == NULL)
 		return -ENOMEM;
